feat: merging different restaurant services
