+++ backend/src/restaurant/restaurant.service.ts
@@ -90,4 +90,25 @@ export class RestaurantService {
       },
     });
   }
+
+  // Fetch restaurants within the specified bounding box
+  async findInBounds(
+    swLat: number,
+    swLng: number,
+    neLat: number,
+    neLng: number,
+  ) {
+    return this.db.restaurant.findMany({
+      where: {
+        latitude: {
+          gte: swLat, // Greater than or equal to swLat
+          lte: neLat, // Less than or equal to neLat
+        },
+        longitude: {
+          gte: swLng, // Greater than or equal to swLng
+          lte: neLng, // Less than or equal to neLng
+        },
+      },
+    });
+  }
 }
