--- backend/src/restaurants/dto/create-restaurant.dto.ts
@@ -1,1 +1,0 @@
-export class CreateRestaurantDto {}
