--- backend/src/restaurants/dto/update-restaurant.dto.ts
@@ -1,4 +1,0 @@
-import { PartialType } from '@nestjs/mapped-types';
-import { CreateRestaurantDto } from './create-restaurant.dto';
-
-export class UpdateRestaurantDto extends PartialType(CreateRestaurantDto) {}
