--- backend/src/restaurants/entities/restaurant.entity.ts
@@ -1,1 +1,0 @@
-export class Restaurant {}
