--- backend/src/restaurants/restaurants.controller.spec.ts
@@ -1,20 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { RestaurantsController } from './restaurants.controller';
-import { RestaurantsService } from './restaurants.service';
-
-describe('RestaurantsController', () => {
-  let controller: RestaurantsController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [RestaurantsController],
-      providers: [RestaurantsService],
-    }).compile();
-
-    controller = module.get<RestaurantsController>(RestaurantsController);
-  });
-
-  it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
