--- backend/src/restaurants/restaurants.controller.ts
@@ -1,58 +1,0 @@
-import {
-  BadRequestException,
-  Controller,
-  Get,
-  Param,
-  Query,
-} from '@nestjs/common';
-import { RestaurantsService } from './restaurants.service';
-import { Restaurant } from 'generated/prisma';
-@Controller('restaurants')
-export class RestaurantsController {
-  constructor(private readonly restaurantsService: RestaurantsService) {}
-
-  @Get()
-  findAll() {
-    return this.restaurantsService.findAll();
-  }
-
-  @Get(':id')
-  findOne(@Param('id') id: string) {
-    return this.restaurantsService.findOne(+id);
-  }
-
-  @Get()
-  findAllWithinBounds(
-    @Query('swLat') swLat: string,
-    @Query('swLng') swLng: string,
-    @Query('neLat') neLat: string,
-    @Query('neLng') neLng: string,
-  ): Promise<Restaurant[]> {
-    if (swLat && swLng && neLat && neLng) {
-      const swLatNum = parseFloat(swLat);
-      const swLngNum = parseFloat(swLng);
-      const neLatNum = parseFloat(neLat);
-      const neLngNum = parseFloat(neLng);
-
-      // Breaking the validation checks across multiple lines for better formatting
-      if (
-        isNaN(swLatNum) ||
-        isNaN(swLngNum) ||
-        isNaN(neLatNum) ||
-        isNaN(neLngNum)
-      ) {
-        throw new BadRequestException('Invalid latitude or longitude values.');
-      }
-
-      return this.restaurantsService.findInBounds(
-        swLatNum,
-        swLngNum,
-        neLatNum,
-        neLngNum,
-      );
-    }
-
-    // If no bounds are provided, return all restaurants
-    return this.restaurantsService.findAll();
-  }
-}
