--- backend/src/restaurants/restaurants.module.ts
@@ -1,9 +1,0 @@
-import { Module } from '@nestjs/common';
-import { RestaurantsService } from './restaurants.service';
-import { RestaurantsController } from './restaurants.controller';
-
-@Module({
-  controllers: [RestaurantsController],
-  providers: [RestaurantsService],
-})
-export class RestaurantsModule {}
