+++ frontend/macos/Flutter/GeneratedPluginRegistrant.swift
@@ -6,11 +6,9 @@ import FlutterMacOS
 import Foundation
 
 import flutter_secure_storage_macos
-import geolocator_apple
 import google_sign_in_ios
 
 func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
   FlutterSecureStoragePlugin.register(with: registry.registrar(forPlugin: "FlutterSecureStoragePlugin"))
-  GeolocatorPlugin.register(with: registry.registrar(forPlugin: "GeolocatorPlugin"))
   FLTGoogleSignInPlugin.register(with: registry.registrar(forPlugin: "FLTGoogleSignInPlugin"))
 }
